requirements.txt
selenium
Django==2.2.6
django-cors-headers==3.1.1
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
Pillow==6.2.1
beautifulsoup4==4.8.1
django-filter==2.2.0
mysqlclient==1.4.4
django-rest-auth==0.9.5